chore(openapi): consolidate admin tags and tighten lint gate (#71)
* chore(openapi): consolidate admin tags, sort schemas, strict lint gate

- Rename admin tags to uniform Admin* prefix (AdminTrustFrameworks,
  AdminSchemaValidation, AdminGraph); merge GraphAdmin + GraphDatabaseAdmin
- Move public /trust-frameworks off Compliance tag to new TrustFrameworks tag;
  extract TrustFrameworksService from ComplianceCheckService
- Merge GraphDatabaseAdminService into GraphAdminService (single delegate
  for the unified AdminGraph tag)
- Sort components.schemas alphabetically; reuse warnings property via YAML
  anchor; reference QueryLanguage enum from QueryInfo (was orphan)
- Tighten Redocly ruleset (operation-summary, security-defined,
  operation-tag-defined, operation-operationId-unique, path-parameters-defined
  as errors); add missing security/summary stanzas; drop unused params
- Lint runs in Maven validate phase via exec-maven-plugin

* fix ci build with sandboxed nodejs

* switch from npx to local installation

* add redocly.yaml lintig config file

* revert to npx

* revert example credential

* cleanup
