+++ openapi/redocly.yaml
@@ -1,0 +1,28 @@
+extends:
+  - minimal
+
+rules:
+  # Hard gates — broken spec / inconsistent operations
+  no-unresolved-refs: error
+  no-identical-paths: warn
+  no-ambiguous-paths: warn
+  operation-operationId-unique: error
+  operation-summary: error
+  operation-tag-defined: error
+  security-defined: error
+  path-parameters-defined: error
+
+  # Hygiene — flag, don't fail
+  no-unused-components: warn
+
+  # Deliberately disabled
+  operation-2xx-response: off
+  operation-4xx-response: off
+  tag-description: off
+  info-license: off
+  info-contact: off
+  no-empty-servers: off
+  no-server-example.com: off
+  no-server-trailing-slash: off
+  operation-description: off
+  parameter-description: off
